pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-google-vertexai>=3.1.1",
    "langchain-huggingface>=1.1.0",
    "langchain[anthropic,google-genai,openai]>=1.1.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
]
